+++ src/sentry_clj/async.clj
@@ -1,102 +1,148 @@
 (ns sentry-clj.async
-  (:require [clojure.string :as string]
-            [clojure.tools.logging :as log]
-            [raven-clj.core :as sentry]
-            [raven-clj.interfaces :as interfaces])
+  (:require [clojure.string :as clj-string]
+            [clojure.tools.logging :as clj-logging]
+            [raven-clj.core :as raven]
+            [raven-clj.interfaces :as raven-interfaces])
   (:import [clojure.lang ExceptionInfo]
            [com.google.common.util.concurrent ThreadFactoryBuilder]
            [com.google.common.util.concurrent MoreExecutors]
+           [java.util UUID]
            [java.util.concurrent ArrayBlockingQueue]
            [java.util.concurrent ExecutorService]
-           [java.util.concurrent RejectedExecutionHandler]
            [java.util.concurrent ThreadPoolExecutor]
-           [java.util.concurrent ThreadFactory]
            [java.util.concurrent ThreadPoolExecutor$DiscardPolicy]
            [java.util.concurrent TimeUnit]))
 
-(declare report-error)
+(defn- error->message
+  [error]
+  (when (instance? ExceptionInfo error)
+    ["\nData:" (ex-data error)]))
 
-(def ^:private sentry-uncaught-exception-handler
+(defn- make-uncaught-exception-handler
+  [logger-fn]
   (reify Thread$UncaughtExceptionHandler
-    (uncaughtException [_ _ exception]
-      (report-error exception "An uncaught error in Sentry worker"))))
+    (uncaughtException [_this _thread exception]
+      (logger-fn :error exception "an uncaught exception occurred while allocating executor worker"))))
 
-(defn- ^ExecutorService create-thread-pool
-  ([worker-count queue-size ^ThreadFactory thread-factory ^RejectedExecutionHandler rejection-policy]
-   (ThreadPoolExecutor. (int worker-count)
-                        (int worker-count)
-                        0
-                        TimeUnit/MILLISECONDS
-                        (ArrayBlockingQueue. (int queue-size))
-                        thread-factory
-                        rejection-policy)))
-
-(defn- ^ExecutorService create-sync-executor []
-  (MoreExecutors/newDirectExecutorService))
-
-(defn- ^ThreadFactory thread-factory [name-format]
+(defn- make-thread-factory
+  [thread-name-format logger-fn]
   (-> (ThreadFactoryBuilder.)
       (.setDaemon true)
-      (.setNameFormat name-format)
-      (.setUncaughtExceptionHandler sentry-uncaught-exception-handler)
+      (.setNameFormat thread-name-format)
+      (.setUncaughtExceptionHandler (make-uncaught-exception-handler logger-fn))
       (.build)))
 
-(defn- ^ExecutorService start-workers [{:keys [sync? worker-count queue-size]}]
-  (log/info "Creating Sentry workers")
-  (if sync?
-    (create-sync-executor)
-    (create-thread-pool worker-count queue-size
-                        (thread-factory "sentry-worker-%d")
-                        (ThreadPoolExecutor$DiscardPolicy.))))
-(defn sentry-report
-  [{:keys [workers config]} level ^Throwable error & msgs]
-  (let [{:keys [enabled dsn env app-name]} config]
-    (when enabled
-      (let [message        (string/join " " (if (instance? ExceptionInfo error)
-                                              (concat msgs ["\nData:" (ex-data error)])
-                                              msgs))
-            event          {:message     message
-                            :level       (name level)
-                            :environment (name env)}
-            ^Runnable task #(try
-                              (sentry/capture dsn
-                                              (interfaces/stacktrace event error [app-name]))
-                              (catch Exception e
-                                (log/warn e "Error while reporting error to sentry")))]
-        (.submit ^ExecutorService workers task)))))
+(defn- make-work-queue
+  [queue-size]
+  (ArrayBlockingQueue. queue-size))
 
-(defmacro logp [level ^Throwable error & msgs]
-  `(if (instance? ExceptionInfo ~error)
-     (log/logp ~level ~error ~@msgs (str "\nData: " (ex-data ~error)))
-     (log/logp ~level ~error ~@msgs)))
+(defn- make-rejected-execution-handler
+  []
+  (ThreadPoolExecutor$DiscardPolicy.))
 
-(defmacro report-error
-  [sentry-reporter ^Throwable error & msgs]
-  `(do (logp :error ~error ~@msgs)
-       (sentry-report ~sentry-reporter :error ~error ~@msgs)))
+(defn- ^ExecutorService make-executor
+  [owner-id
+   logger-fn
+   {:keys [async?
+           async-executor-workers-idle-count
+           async-executor-workers-maximum-count
+           async-executor-workers-keep-alive-time
+           async-executor-workers-keep-alive-unit
+           async-executor-workers-queue-size
+           async-executor-workers-name-format]
+    :or   {async?                                 false
+           async-executor-workers-idle-count      10
+           async-executor-workers-maximum-count   10
+           async-executor-workers-keep-alive-time 0
+           async-executor-workers-keep-alive-unit TimeUnit/MILLISECONDS
+           async-executor-workers-queue-size      10}}]
+  (if async?
+    (let [async-executor-workers-name-format (or async-executor-workers-name-format
+                                                 (str "corvus-" owner-id "-worker-%d"))]
+      (ThreadPoolExecutor. async-executor-workers-idle-count
+                           async-executor-workers-maximum-count
+                           async-executor-workers-keep-alive-time
+                           async-executor-workers-keep-alive-unit
+                           (make-work-queue async-executor-workers-queue-size)
+                           (make-thread-factory async-executor-workers-name-format logger-fn)
+                           (make-rejected-execution-handler)))
+    (MoreExecutors/newDirectExecutorService)))
 
-(defmacro report-warn
-  [sentry-reporter ^Throwable error & msgs]
-  `(do (logp :warn ~error ~@msgs)
-       (sentry-report ~sentry-reporter :warning ~error ~@msgs)))
+(defn make-captor
+  ([sentry
+    application]
+   (make-captor sentry application {}))
+  ([{:keys [logger-fn]
+     :or   {logger-fn (fn [level
+                           message
+                           & more]
+                        (clj-logging/logp level message more))}
+     :as   sentry}
+    application
+    options]
+   (let [captor-id (UUID/randomUUID)]
+     (logger-fn :info (format "initializing captor[%s] executor" captor-id) "_/\\__/\\__0>")
+     (-> {:id captor-id}
+         (assoc :sentry (assoc sentry :logger-fn logger-fn))
+         (assoc :application application)
+         (assoc :executor (make-executor captor-id logger-fn options))))))
 
-(defn create-reporter [{:keys [enabled dsn env app-name sync? worker-count queue-size thread-termination-wait-s]}]
-  (let [enabled (if (nil? enabled) true enabled)
-        config  {:sync?                     (or sync? false)
-                 :enabled                   enabled
-                 :dsn                       dsn
-                 :env                       env
-                 :app-name                  app-name
-                 :worker-count              (or worker-count 10)
-                 :queue-size                (or queue-size 10)
-                 :thread-termination-wait-s (or thread-termination-wait-s 10)}]
-    {:config  config
-     :workers (when enabled
-                (start-workers (select-keys config [:sync? :worker-count :queue-size])))}))
+(defn kill-captor
+  ([captor]
+   (kill-captor captor {}))
+  ([{captor-id :id
+     :keys     [sentry
+                executor]}
+    {:keys [executor-await-termination-time
+            executor-await-termination-unit]
+     :or   {executor-await-termination-time 10
+            executor-await-termination-unit TimeUnit/SECONDS}}]
+   (let [logger-fn     (get sentry :logger-fn)
+         logger-action (format "terminating captor[%s] executor" captor-id)]
+     (logger-fn :info logger-action "(-.-)Zzz...")
+     (try
+       (.shutdown ^ExecutorService executor)
+       (let [terminated? (.awaitTermination
+                           ^ExecutorService executor
+                           executor-await-termination-time
+                           executor-await-termination-unit)]
+         (if terminated?
+           (logger-fn :info "success in" logger-action)
+           (logger-fn :error "failed in" logger-action))
+         terminated?)
+       (catch Exception error
+         (logger-fn :error error "failed in" logger-action))))))
 
-(defn shutdown-reporter
-  [{:keys [config workers]}]
-  (when workers
-    (.shutdown ^ExecutorService workers)
-    (.awaitTermination ^ExecutorService workers (:thread-termination-wait-s config) TimeUnit/SECONDS)
-    (log/info "stopped Sentry workers")))
+(defn capture!
+  [{captor-id :id
+    :keys     [sentry
+               application
+               executor]}
+   level
+   error
+   & messages]
+  (let [event {:level       (-> level
+                                (or :error)
+                                name)
+               :environment (-> application
+                                (get :environment)
+                                name)
+               :message     (->> error
+                                 error->message
+                                 (concat messages)
+                                 (clj-string/join " "))}]
+    (.submit ^ExecutorService executor
+             ^Runnable (fn []
+                         (let [logger-fn     (get sentry :logger-fn)
+                               logger-action (format "capturing event %s with captor[%s]"
+                                                     event
+                                                     captor-id)]
+                           (try
+                             (raven/capture
+                               (get sentry :data-source-name)
+                               (raven-interfaces/stacktrace
+                                 event
+                                 error
+                                 [(get application :name)]))
+                             (catch Exception error
+                               (logger-fn :error error "error while" logger-action))))))))